# docs/api/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    ApiProblem:
      description: RFC 9457 problem detail with stable application code
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        code:
          type: string
        requestId:
          type: string
...
    CreateInvitationRequest:
      type: object
      required:
      - actorUserId
      - expiresAt
      properties:
        actorUserId:
          type: string
          pattern: \S
        expiresAt:
          $ref: "#/components/schemas/Instant"
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    RedeemInvitationRequest:
      type: object
      required:
      - invitationToken
      - username
      - password
      properties:
        invitationToken:
          type: string
          pattern: \S
        username:
          type: string
          pattern: \S
        password:
          type: string
          pattern: \S
    RevokeInvitationRequest:
      type: object
      required:
      - actorUserId
      properties:
        actorUserId:
          type: string
          pattern: \S
...
    SessionLoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          pattern: \S
        password:
          type: string
          pattern: \S
    SessionLoginResponse:
      type: object
      properties:
        sessionId:
          type: string
        token:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
info:
  description: Private messaging backend API
  title: ChatBackend API
  version: 0.7.0-SNAPSHOT
security:
- bearerAuth: []
paths:
...
  /api/v1/invitations:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Create Invitation
      tags:
      - Invitation Resource
  /api/v1/invitations/redeem:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemInvitationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      security: []
      summary: Redeem Invitation
      tags:
      - Invitation Resource
  /api/v1/invitations/{invitationId}/revoke:
    post:
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeInvitationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Revoke Invitation
      tags:
      - Invitation Resource
...
  /api/v1/sessions:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionLoginRequest"
        required: true
      responses:
        "200":
          description: Authenticated session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionLoginResponse"
        "400":
          description: Malformed or invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: Invalid credentials
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "429":
          description: Authentication attempt limit exhausted
          headers:
            Retry-After:
              description: Seconds until another attempt is permitted
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "500":
          description: Authentication service unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      security: []
      summary: Login
      tags:
      - Session Resource
...
  /api/v1/sessions/users/{userId}/revoke-all:
    post:
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Revoke All Sessions
      tags:
      - Session Resource
...
servers:
- url: http://127.0.0.1:8080
  description: Auto generated value